added cards
