+++ src/pages/Students/ZStudentsPage.css
@@ -7,3 +7,6 @@
   padding-left: 94px;
   padding-right: 30px;
 }
+
+
+
