+++ src/pages/Templates/ZTemplatesPage.css
@@ -7,3 +7,14 @@
   padding-left: 94px;
   padding-right: 30px;
 }
+
+.flex-templates {
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  padding-left: 94px;
+  padding-right: 30px;
+
+}
